Attempt to fix the Keep Save Files option.
Soemtimes the save file gets deleted when I die even if I have the
Keep Save Files option checked.  I haven't figured out why that happens,
so change the default to be that deleteGame doesn't delete the file if
that option is set, unless a new "force" option is specified, and only
pass "force = true" in the cases where the file should always be deleted
(mostly when ending the game after getting the amulet).